scripts/classify.py
import numpy as np
import pandas as pd
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics import average_precision_score

def run():
	df = pd.read_csv('./resources/feature_dataset.csv', sep=',')
	print(df)