requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
diff-match-patch==20121119
Django==2.0
django-extensions==2.0.5
django-import-export==1.0.0
et-xmlfile==1.0.1
idna==2.5
jdcal==1.3
lyricsgenius==0.4.1
numpy==1.14.2
odfpy==1.3.6
openpyxl==2.5.1
pandas==0.22.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
requests==2.18.1
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
spotipy==2.4.4
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.21.1
xlrd==1.1.0
xlwt==1.3.0
